+++ src/give/PaymentMethod.js
@@ -4,20 +4,19 @@ import Header from '@ui/Header';
 import BackgroundView from '@ui/BackgroundView';
 import PaddedView from '@ui/PaddedView';
 import EditSavedPaymentMethodForm from '@ui/forms/EditSavedPaymentMethodForm';
-import DeleteSavedPaymentMethodButton from '@ui/forms/DeleteSavedPaymentMethodButton';
 import { compose, mapProps } from 'recompose';
-import Spacer from '@ui/Spacer';
+
+import { Title } from './styles';
 
 export function PaymentMethod({ id } = {}) {
   return (
     <BackgroundView>
       <Header titleText="Payment Method" />
-      <KeyboardAwareScrollView>
+      <KeyboardAwareScrollView enableOnAndroid>
         <PaddedView>
-          <EditSavedPaymentMethodForm id={id} />
-          <Spacer />
-          <DeleteSavedPaymentMethodButton id={id} />
+          <Title>Edit Account</Title>
         </PaddedView>
+        <EditSavedPaymentMethodForm id={id} />
       </KeyboardAwareScrollView>
     </BackgroundView>
   );
